se agrego toda la logica CRUD para ejercicios

+++ src/main/java/com/GymMaster/backendGym/entity/Progreso.java
@@ -29,5 +29,5 @@ public class Progreso {
     private Integer duracion;
 
     @Column (name = "peso")
-    private Integer peso;
+    private Double peso;
 }
